requirements.txt
# FastAPI & Web
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# AI & ML
google-genai==1.5.0
sentence-transformers==3.3.1
torch==2.5.1
torchvision==0.20.1

# OCR
easyocr==1.7.2

# Vector DB
chromadb==0.6.3

# PDF & QR
reportlab==4.2.5
qrcode[pil]==8.0

# Image Processing
Pillow==11.1.0

# S3 Storage
boto3==1.36.2

# Redis
redis==5.2.1

# Utilities
numpy==2.2.1
pandas==2.2.3
